changed who identifies the type
Originally I was thinking about finding out the Type of a Variable at Tokenization time, but it's actually the Parser's job. ("(DEFINE) (any DEFINE DESCRIPTOR) (IDENTIFIER (name of the variable)) (ASSIGN) (NUMBER)(ENDLINE)" Runs into an Error because to create a variable in this language it's supposed to be " (DEFINE) (and DEFINE DESCRIPTOR) (IDENTIFER (Type of the variable)) (IDENTIFIER(Name of the Variable)) (ASSIGN) (NUMBER) (ENDLINE)", so you can't write the Type and the Name together. And now we need whitespaces again :c)